[core] Moved socket group code to separate files
